move highlighting to utils
