+++ src/utils/highlighting.js
